+++ Changelog.md
@@ -9,3 +9,7 @@ Initial release.
 - Specify that `ObjectPool<T>.Return(T obj)` must never fail silently if `obj` is `null`.
 - Fix `DynamicObjectPool<T>` not auto-trimming.
 - Fix `ThreadLocalOverPerLockedStacksObjectPool<T>.Rent()` (implementation behind `ObjectPool<T>.Shared.Rent()`) throwing when getting elements from global reserve.
+
+##
+- Remove `class` constraint in generic parameter `T` in `ObjectPool<T>`.
+- Add `DynamicValueObjectPool<T>`.